Reorganized them Colors
Added more static colors to reference in ZColors, edited a few comments, added skeleton code for game things

+++ src/main/java/zes/core/engine/utils/ZColor.java
@@ -2,12 +2,12 @@ package zes.core.engine.utils;
 
 public class ZColor {
 	private String colorName;
-	private int r, g, b, a;
+	private float r, g, b, a;
 	
 	// Constructors
 	public ZColor(String colorIn) { this(colorIn, 0, 0, 0, 0); }
-	public ZColor(String colorIn, int rIn, int gIn, int bIn) { this(colorIn, rIn, gIn, bIn, 0); }
-	public ZColor(String colorIn, int rIn, int gIn, int bIn, int aIn) {
+	public ZColor(String colorIn, float rIn, float gIn, float bIn) { this(colorIn, rIn, gIn, bIn, 0); }
+	public ZColor(String colorIn, float rIn, float gIn, float bIn, float aIn) {
 		colorName = colorIn;
 		r = rIn;
 		g = gIn;
@@ -15,19 +15,19 @@ public class ZColor {
 		a = aIn;
 	}
 	
-	public void setAll(int rIn, int gIn, int bIn) { setAll(rIn, gIn, bIn); }
-	public void setAll(int rIn, int gIn, int bIn, int aIn) { r = rIn; g = gIn; b = bIn; a = aIn; }
+	public void setAll(float rIn, float gIn, float bIn) { setAll(rIn, gIn, bIn); }
+	public void setAll(float rIn, float gIn, float bIn, float aIn) { r = rIn; g = gIn; b = bIn; a = aIn; }
 	
 	// Getters
 	public String getColorName() { return colorName; }
-	public int getR() { return r; }
-	public int getG() { return g; }
-	public int getB() { return b; }
-	public int getA() { return a; }
+	public float getR() { return r; }
+	public float getG() { return g; }
+	public float getB() { return b; }
+	public float getA() { return a; }
 	
 	// Setters
-	public void setR(int rIn) { r = rIn; }
-	public void setG(int gIn) { g = gIn; }
-	public void setB(int bIn) { b = bIn; }
-	public void setA(int aIn) { a = aIn; }
+	public void setR(float rIn) { r = rIn; }
+	public void setG(float gIn) { g = gIn; }
+	public void setB(float bIn) { b = bIn; }
+	public void setA(float aIn) { a = aIn; }
 }
